Update battle_calculator.php5
I changed the code because I still think the *2 is the problem. If it is the problem then the bug will now be fixed, if it is not then nothing will change.

I added two separate get creature capture functions, one for attackers and one for defenders and then I added the *2 directly to the defenders int. I do not think you can multiply a FleetData object by a number.

Further I looked thoroughly through the code and there is no problem on or around line 170, its a bit confusing at first but I am positive everything works fine there. 

And just to reiterate the bug we are trying to correct makes ALL attacking creatures immune to capture whether the attacker has FBI or not. I'm not sure when this bug came around but it may have been before you commented out line 170.